MDL-54724 qtype_multianswer: avoid fatal error during restore
During a course restore, if a new question id is not found for the id's in the
sequence, we end up with empty values in the sequence list.

This change filters the sequence list of empty values after the list has been
built (via get_mappingid) to ensure no empty values are present.