MDL-72229 core_files: Handle gzip steams in tgz_extractor
If the underlying file_system provides access to gzip files via a stream with a
stream filter attached to handle the decompression - gzread() can return less
bytes than asked for after php 7.4.15.

This change handles this by reading from the file handle again until there is
enough bytes in the buffer to meet the amount required for a tar block.