MDL-76778 cache: Provide static cache via cache_application
Without this, the standard features of the `cache` class, namely
versioning, are not provided by the cache instances generated thanks to
allow_temporary_caches.

With this, we instantiate a cache_application, backed by a static cache.

End-result is we still have a static cache - but it has all the extra
features one expects to have when interacting with caches.